[WIP] Move contents of Web3Helper to utils.py, refactor to module funcs
