+++ ocean_lib/web3_internal/constants.py
@@ -9,3 +9,13 @@ GAS_LIMIT_DEFAULT = 1000000
 MIN_GAS_PRICE = 1000000000
 
 ZERO_ADDRESS = "0x0000000000000000000000000000000000000000"
+
+DEFAULT_NETWORK_NAME = "ganache"
+NETWORK_NAME_MAP = {
+    1: "Mainnet",
+    2: "Morden",
+    3: "Ropsten",
+    4: "Rinkeby",
+    42: "Kovan",
+    100: "xDai",
+}
